ajout power up mode

+++ src/game/Mode/PowerUp.hpp
@@ -14,7 +14,7 @@ class PowerUp: public virtual Mode {
 			
 	public:
 		
-		PowerUp(unsigned);
+		PowerUp(long);
 		virtual ~PowerUp();
 		virtual void start(Grid *,Grid*, Stopper_Thread*)override;		
 		virtual void init_game(bool is_player)override;
